[1/2] sqlite query perf: try to continue native sqlite statements
If we avoid resetting statements, we may be able to continue where we
left off last time. This saves us the work of repeating the whole query
over and over again when we just want to fill sequential windows.

To achieve this, add the SQLiteContinuation class, which holds state
usable by SQLiteConnection. Each statement in the connection's cache
will reference the last continuation, and if an incoming continuation
matches, we will continue the statement where we left off.

When continuing, SQLiteConnection.executeForCursorWindow() now takes
startPos more lightly; if a continuation is possible, actualPos may
be clamped to the current statement position (in addition to the old
behavior of actualPos moving up to make sure that requiredPos fits).

However, leaving statements alive ("busy", in sqlite terms) may cause
some operations to return SQLITE_LOCKED (example: DROP TABLE). So, we
have to add some logic to reset any busy statements whenever a
statement fails with an SQLiteTableLockedException.

Leaving statements alive also keeps sqlite's internal query state,
increasing memory usage. To mitigate this, SQLiteQuery makes a call
to discontinue(), informing the lower layers when the continuation is
no longer active. If the owning connection is taken by another thread
at that point, the discontinuation is queued, and handled when the
connection is returned to the pool.

My benchmark - moveToNext() through a 25000-row result set, with the
window being able to fit around 6700 - improved from 1750ms to 750ms.

Future work:
 * Two queries with identical SQL strings could break continuation for
   each other (fighting for the same PreparedStatement). This could
   perhaps be solved by moving statements out of the cache (to another
   structure) while they're continuable.
 * Consider the continuation's connection id in waitForConnection().

Change-Id: I30fe9a72ef7c08cc5adcf233cca2ae216b97eb86