Revert "Force Expanded Notifications (1/2)"
This causes a stack of notifications (like Inbox or Telegram) to jitter
when first opened; I suspect it is because the system has to take a
small amount of time to calculate if it has to expand the notifications
after the stack is expanded. Removing this commit eliminates that jank.

This reverts commit 227185421ca41e446b937512830cc5a918accbcb.